// resources/js/game/components/actions/partials/floating-cards/crafting-section/sections/batch-crafting/enums/batch-crafting-disposition.ts
export enum BatchCraftingDisposition {
  KEEP = 'keep',
  SELL = 'sell',
  DESTROY = 'destroy',
  LIST = 'list',
  DISENCHANT = 'disenchant',
  USE_NOW = 'use_now',
  KEEP_BEST_SELL_REST = 'keep_best_sell_rest',
  KEEP_BEST_DESTROY_REST = 'keep_best_destroy_rest',
  KEEP_BEST_DISENCHANT_REST = 'keep_best_disenchant_rest',
}