// resources/js/game/components/actions/partials/floating-cards/crafting-section/sections/batch-crafting/types/craft-set-output-selection.ts
import { BatchCraftingDisposition } from '../enums/batch-crafting-disposition';
import { BatchCraftingOutputDestination } from '../enums/batch-crafting-output-destination';

export default interface CraftSetOutputSelection {
  disposition: BatchCraftingDisposition;
  output_destination: BatchCraftingOutputDestination | null;
  output_set_id: number | null;
  output_set_name: string | null;
}